src/test/java/io/github/vlad324/n1275/SolutionTest.java
package io.github.vlad324.n1275;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

class SolutionTest {

    private final Solution solution = new Solution();

    private static Stream<Arguments> testDataProvider() {
        return Stream.<Arguments>builder()
            .add(arguments(new int[][]{{0, 0}, {2, 0}, {1, 1}, {2, 1}, {2, 2}}, "A"))
            .add(arguments(new int[][]{{0, 0}, {1, 1}, {0, 1}, {0, 2}, {1, 0}, {2, 0}}, "B"))
            .add(arguments(new int[][]{{0, 0}, {1, 1}, {2, 0}, {1, 0}, {1, 2}, {2, 1}, {0, 1}, {0, 2}, {2, 2}}, "Draw"))
            .add(arguments(new int[][]{{0, 0}, {1, 1}}, "Pending"))
            .build();
    }

    @ParameterizedTest
    @MethodSource("testDataProvider")
    void should_produce_expected_result(int[][] moves, String expected) {
        // when
        final var actual = solution.tictactoe(moves);

        // then
        assertThat(actual).isEqualTo(expected);
    }

}